+++ lib/frontend/tlp-popovers/src/popovers.ts
@@ -24,6 +24,10 @@ export const POPOVER_SHOWN_CLASS_NAME = "tlp-popover-shown";
 export const EVENT_TLP_POPOVER_HIDDEN = "tlp-popover-hidden";
 export const EVENT_TLP_POPOVER_SHOWN = "tlp-popover-shown";
 
+export const EVENT_POPOVER_FORCE_CLOSE = "tlp-popover-force-close";
+
+const TIMEOUT_BEFORE_POPOVER_IS_HIDDEN_ON_MOUSEOUT_IN_MS = 200;
+
 type Configuration = {
     readonly anchor: HTMLElement;
     readonly trigger: Trigger;
@@ -191,7 +195,16 @@ function getConfiguration(popover_trigger: HTMLElement, options: PopoverOptions)
     };
 }
 
-type EventType = "click" | "mouseout" | "mouseover" | "keyup" | "focus" | "blur";
+type EventType =
+    | "click"
+    | "mouseout"
+    | "mouseenter"
+    | "mouseleave"
+    | "mouseover"
+    | "keyup"
+    | "focus"
+    | "blur"
+    | typeof EVENT_POPOVER_FORCE_CLOSE;
 
 interface EventListener {
     element: EventTarget;
@@ -211,12 +224,16 @@ function buildListeners(
         return [
             buildFocusListener(doc, popover_trigger, popover_content, updatePositionOfContent),
             buildBlurListener(doc, popover_trigger, popover_content),
+            buildForceClosePopover(doc, popover_content),
         ];
     }
     if (configuration.trigger === "hover") {
         return [
             buildMouseOverListener(doc, popover_trigger, popover_content, updatePositionOfContent),
             buildMouseOutListener(doc, popover_trigger, popover_content),
+            buildMouseEnterListener(doc, popover_content),
+            buildMouseLeaveListener(doc, popover_content),
+            buildForceClosePopover(doc, popover_content),
         ];
     }
     if (configuration.trigger === "click") {
@@ -229,6 +246,7 @@ function buildListeners(
             ),
             buildDocumentClickListener(doc, popover_trigger, popover_content),
             buildEscapeListener(doc, popover_content),
+            buildForceClosePopover(doc, popover_content),
         ];
         for (const dismiss of dismiss_buttons) {
             listeners.push(buildDismissClickListener(doc, dismiss));
@@ -251,6 +269,20 @@ function destroyListeners(listeners: EventListener[]): void {
     }
 }
 
+function buildForceClosePopover(doc: Document, popover_content: HTMLElement): EventListener {
+    return {
+        element: doc,
+        type: EVENT_POPOVER_FORCE_CLOSE,
+        handler(): void {
+            if (!popover_content.classList.contains(POPOVER_SHOWN_CLASS_NAME)) {
+                return;
+            }
+
+            hidePopover(popover_content);
+        },
+    };
+}
+
 function buildFocusListener(
     doc: Document,
     popover_trigger: HTMLElement,
@@ -266,6 +298,8 @@ function buildFocusListener(
         },
     };
 }
+
+let timeout = 0;
 
 function buildBlurListener(
     doc: Document,
@@ -307,10 +341,43 @@ function buildMouseOutListener(
         element: popover_trigger,
         type: "mouseout",
         handler(): void {
-            hideAllShownPopovers(doc);
-            hidePopover(popover_content);
+            startTimeout(doc, popover_content);
         },
     };
+}
+
+function buildMouseLeaveListener(doc: Document, popover_content: HTMLElement): EventListener {
+    return {
+        element: popover_content,
+        type: "mouseleave",
+        handler(): void {
+            startTimeout(doc, popover_content);
+        },
+    };
+}
+
+function buildMouseEnterListener(doc: Document, popover_content: HTMLElement): EventListener {
+    return {
+        element: popover_content,
+        type: "mouseenter",
+        handler(): void {
+            stopTimeout();
+        },
+    };
+}
+
+function startTimeout(doc: Document, popover_content: HTMLElement): void {
+    timeout = window.setTimeout(() => {
+        hideAllShownPopovers(doc);
+        hidePopover(popover_content);
+    }, TIMEOUT_BEFORE_POPOVER_IS_HIDDEN_ON_MOUSEOUT_IN_MS);
+}
+
+function stopTimeout(): void {
+    if (timeout) {
+        window.clearTimeout(timeout);
+        timeout = 0;
+    }
 }
 
 function buildTriggerClickListener(
@@ -387,9 +454,9 @@ function hidePopover(popover_content: Element): void {
 }
 
 function hideAllShownPopovers(doc: Document): void {
-    for (const popover of doc.querySelectorAll("." + POPOVER_SHOWN_CLASS_NAME)) {
-        hidePopover(popover);
-    }
+    doc.dispatchEvent(new CustomEvent(EVENT_POPOVER_FORCE_CLOSE));
+
+    stopTimeout();
 }
 
 function showPopover(popover_content: HTMLElement, updatePositionOfContent: () => void): void {