+++ lib/frontend/project-sidebar/CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## In development
 
-* In order to not clutter the sidebar when the number of linked projects is too big,
+* 🚸 In order to not clutter the sidebar when the number of linked projects is too big,
   a new configuration variable is added: `project.linked_projects.nb_max_projects_before_popover`.
   When the number of linked projects exceed this limit, then they are no more displayed in the sidebar,
   only in the popover on hover. Default is `5`.
-* Bump vue: 3.2.37 -> 3.3.4
+* 🐛 Linked projects displayed in the popover are now clickable links
+* ⬆️ Bump vue: 3.2.37 -> 3.3.4
 
 ## 2.3.0